api/requirements.txt
Django==3.1.3
django-cors-headers==3.2.1
djangorestframework==3.12.2
psycopg2>=2.7,<3.0
requests==2.18.4
